More smarter collection's events bubbling
First of all a new event born, the "*". Every store and collection
has this event now. It's triggered together with every other event, and
your first argument is the name of the event that triggered the wildcard
event.

With this, the collection store doesn't trigger "change" anymore when
some children stores change, this event was moved into the "store:change"
event, but now, every event on the children stores is triggered on
"store:<name-of-event>" event, giving much more flexibility to listen
children stores events on the collection.