Adjusment on Fluxo.ObjectStore#reset
Lets us say that you have the following class:

class Foo extends Fluxo.ObjectStore {
  static defaults = {
    emails: []
  };

  static computed = {
    hasEmails: ["change:emails"]
  };

  hasEmails () {
    return this.data.hasEmails.length > 0;
  }
};

With the past implementation of #reset the #clear method calling going to clear
all attributes and #hasEmails will be computed with `this.data.hasEmails` as
undefined and it could break some computed properties that aren't prepared
to deal with unexpected undefined.

This adjustment makes sure to be more smart about this situation.