Releasing (or not) nested stores
Now, with the nested events bubbling setup we pledged with this nested kind of use
and nested stores could being stale when unset or swap this attribute. For this
reason Fluxo will try whenever it's possible release nested stores when it
goes out of its parent.

Important: it's opt-out (case you want continue using some nested instance out of
its parent) using the `releaseNested: false` option on the methods:

* `Fluxo.ObjectStore#set`
* `Fluxo.ObjectStore#setAttribute`
* `Fluxo.ObjectStore#reset`
* `Fluxo.ObjectStore#clear`
* `Fluxo.ObjectStore#unsetAttribute`