Computed properties and subsets on `change` event
This commit allows the user to delegate the update of its computed properties and
subsets on the `change` event without triggering the `change:<subset-or-computed-property-name>`
after the `change` event. To this, the developer must leave the dependent events array
empty and Fluxo will know that these computed things must be called when a
"change event cycle" is finishing.

It also will warn the developer about this if he puts the "change" event
as a dependent event to recompute the computed things.