Reverting the need of declaring the change event for subsets and computed property
Declaring an empty array as dependent events to recompute the subsets and computed
properties can mislead to newcomers since it does not show in a clearer way the
dependent event, so, the developer must declare the `change` event.

An important note: the developer shouldn't declare any other dependent event
with `change` since that `change` event is the end of the line of "state mutation
cycle".