Removes unnecessary code at the change events
There's no need to trigger the events attached on 'change' event before trigger it.

This was added to avoid ifinite loop when subsets and computed change when it's
attached to the change event, but today the computed and subsets doesn't trigger
the change event since it's always recomputed from other change event.